Rename extension to Logging Manager
