+++ deployment/src/main/java/io/quarkiverse/loggingmanager/quarkus/logging/manager/deployment/LoggingManagerBuildItem.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.loggingui.quarkus.logging.ui.deployment;
+package io.quarkiverse.loggingmanager.quarkus.logging.manager.deployment;
 
 import io.quarkus.builder.item.SimpleBuildItem;
