+++ deployment/src/main/java/io/quarkiverse/loggingmanager/quarkus/logging/manager/deployment/LoggingManagerOpenAPIFilter.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.loggingui.quarkus.logging.ui.deployment;
+package io.quarkiverse.loggingmanager.quarkus.logging.manager.deployment;
 
 import java.util.Collections;
 import java.util.HashMap;
