+++ deployment/src/main/java/io/quarkiverse/loggingmanager/quarkus/logging/manager/deployment/LoggingManagerUIConfig.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.loggingui.quarkus.logging.ui.deployment;
+package io.quarkiverse.loggingmanager.quarkus.logging.manager.deployment;
 
 import io.quarkus.runtime.annotations.ConfigGroup;
 import io.quarkus.runtime.annotations.ConfigItem;
