+++ deployment/src/test/java/io/quarkiverse/loggingmanager/quarkus/logging/manager/deployment/CustomEndpointTest.java
