+++ runtime/src/main/java/io/quarkiverse/loggingmanager/quarkus/logging/manager/LevelHandler.java
