+++ ui/src/main/webapp/logging-manager/index.html
