+++ ui/src/main/webapp/logging-manager/loggingmanager.js
