+++ ui/src/main/webapp/logging-manager/style.css
